developer installation from git
Please have a look if this works as intended. 

I currently use the enviroment made by conda, as this is the easiest way. It would however be nice if there is an `environment.yaml`, as there would be no dependency issues if between the master and develop branches.

P.S. I moved the Python requirement to the TODO section, as it seemed misplaced.
P.P.S. also added the config option we seem to need to install the latest version without specifying the version number.